Item#13, IMPORTANT! Changes 'did exist in CRM' logic.
When we can't find a thing in CRM, decision of whether to either a)
delete it locally, or b) add it to CRM WAS taken on the basis of
(pseudocode):

  var didExistOnCRM = !string.IsNullOrEmpty(syncState.OModifiedDate.ToString())
  if (didExistOnCRM)
    DeleteLocally(syncState);
  else
    AddToCrm(syncState);

Trouble is that DateTime.ToString() always returns a non-empty, non-null
string!

So changed the logic to:

  var didExistOnCRM = !string.IsNullOrEmpty(syncState.CrmEntityId)

Honestly this seems rather obvious, so testing will tell whether there's
something subtly wrong with it. (But the previous version was obviously,
grossly wrong.)