Task#56 Propagates Outlook deletions for appointments
Fixes regression introduced in the 2.2.3 builds: while refactoring the
syncing code and removing redundancy, I'd inadvertantly set
'AppointmentSyncing.PropagatesLocalDeletions' to false instead of true.

Note that issue #56 also mentions that local deletions of Contacts are
not being propagated. I've been unable to reproduce, and want to have
the report confirmed.