#85: Seems like a pretty complete fix
I've got rid of the horrid 'dynamic' objects, and refactored significantly. Everything's a lot cleaner and I believe there are substantially fewer lines of code. There may still be a number of cases where we're casting things to strings and back again where we don't need to, but they would take time to track down.