#112: This is broken, but there is good work in it
Currently, syncing doesn't seem to be working at all, but the logic for syncing is greatly cleaned up.

I've put a lot of sync permission caching stuff in Synchroniser; this is suspect, and in any case should be in its own class in SuiteCRMClient. But this isn't the main cause of the breakage, because synchronisation was already not working before this was added.

TODO: Fix.