#144: fix
The issue here, I believe, was not permissions but threading. Multiple mails were archived simultaneously in separate threads, but using the same directory path, and the directory was deleted as the operation exited. The problem is obvious: one thread deletes the directory while another is still trying to write to it.