#235: Fixed serious bug which prevented appointment syncing
AppointmentSyncStates were created with their crmType taken from the crmType of the underlying Outlook item. The problem was that at the time this happened the property had not yet been fixed up.
This is a long standing bug - from before February - which we'd failed to spot. It affected appointments which were created in Outlook but which had not synced by the end of the Outlook session.