#245: Problem probably server side.
The present code successfully sends requests to create links to each of the modules selected for each email address selected; CRM reports success on all these links. However, in the CRM user interface, the emails do NOT show in the history of entries in the module lag_Lagging_Indicator. I suggest the problem is server side.