4509: Added special purpose exception, and handler for it.
Validated that throwing the exception does not interfere with actually saving the item.

There's more to this issue than this - if we aren't syncing, we cannot in principle check whether or not it's a valid CRM id (I think) and so should not. But that takes some redesign which I'm not up to today.